Change 737 Version ver-3_2_1_1 Aranea 1.2.1
git-svn-id: file:///root/var/svn/aranea/branches/branch-3_1_1@4697 baba0d3d-6527-0410-ba98-ffd8059025de

+++ examples/main/src/org/araneaframework/example/main/web/sample/SampleActionFormWidget.java
@@ -101,7 +101,7 @@ public class SampleActionFormWidget extends TemplateBaseWidget {
   }
 
   protected BigDecimal calculateTotal(BigDecimal price, Integer quantity) {
-    return price.multiply(new BigDecimal(quantity.intValue()));
+    return price.multiply(new BigDecimal(quantity.doubleValue()));
   }
 
   protected BigDecimal calculateVatTotal(BigDecimal total, BigDecimal vat) {
